Some cleanup

+++ ModNetworkAPI/Server.cs
@@ -57,7 +57,7 @@ namespace ModNetworkAPI
         /// <param name="message">Text that will be displayed in client chat</param>
         /// <param name="data">A serialized object to be sent across the network</param>
         /// <param name="isReliable">Enture delivery of the packet</param>
-        public void SendCommandToPlayersInRange(Vector3D point, long radius, string commandString, string message = null, byte[] data = null, bool isReliable = true)
+        public void SendCommandToPlayersInRange(Vector3D point, float radius, string commandString, string message = null, byte[] data = null, bool isReliable = true)
         {
             SendCommandToPlayersInRange(new BoundingSphereD(point, radius), commandString, message, data, isReliable);
         }
